1D Model implemented in Python
This implementation needs still a bit clean up and a clear interface to Julia. I will add a short documentation file in the "docs" folder, giving an example on how to use it.
The implementation is based on impedance transformation and 1D ray tracing (analytical), which should be equivalent to the transfer matrices, but the formulae are a bit reshuffled.